requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.25
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==4.4.0
importlib-resources==5.4.0
Pillow==8.4.0
PyJWT==2.4.0
pytz==2025.2
sqlparse==0.4.4
typing_extensions==4.1.1
zipp==3.6.0
